Cleaned up Code, Projectiles shoot from AlienShip and Score/Lives Displayed
This update I moved AlienShip and Asteroid code and methods to their classes to clean up. Added projectiles shooting from AlienShip to playerShip. I also converted the AlienShip to PVector attibutes and movement. I also introduced a score display, player lives display and a method of the player ship dying so to speak with it flashing until its back to active.