Added last Collision Detection
Added the last two collision detection tests - asteroid to player and bullet to asteroid. Fixed iteration issue that was causing collision detection errors. Also changed part of splitAsteroid method to make it split itself with out a parameter. Added changes from Elsa's branch and changed the sound mechanism to the Minim library as an error was occuring in the console otherwise.